Untangle dir management
Replace use of MkdirAll() with Mkdir() per directory. It is more verbose
and repetetive but makes it easy to control ownership and perms for each
dir or file. Also do not do any implicit mkdir when trying to create a
normal file but instead just expect the directory to already have been
created.