Replace any use of names with IDs
Thinking about this some more, while it is nice to use organisation and
service names in places for easy reference this does bring additional
management overhead as such names could potentially be changed causing
problems like "should we reload a service if a config file changed if the
only thing that changed was a comment mentioning the name".

For this reason lets abandon the use of names in any config files or
paths. Using the immutable ID everywhere looks worse but since we can
depend on them "never changing" it makes all management tasks much
easier.

There is also a security implication here: since users have no influence
over the selection of ID content (the UUID values are assigned by the
database) there is less attack surface for injections into any of the
files we are generating.