Take full control of executed varnishd command
The upstream container entrypoint script will check if the first
argument of Cmd is a flag and if it is the flags in Cmd will just be
appended to an existing set of flags.

For us this meant opening listening ports we do not need, and also
reading in a default.vcl via an intial -f. Since we are setting our own
-f already this resulted in both an unused "boot0" and our content in
"boot1" in "varnishadm vcl.list".

Start Cmd with "varnishd" to effectively take over the whole argument
list ourselves so we just load one vcl file on boot (now called just
"boot" in vcl.list) and listen only to a unix socket. This needs the
addition of "-F" from the entrypoint script for varnishd to stay in the
foreground.