Add support for reloading varnish vcl
If we detect that the active symlink has been updated for a service we
use varnishadm to load and use the new vcl file. We will also discard
unused previous config entries to keep things tidy.

Right now we only look for vcl names matching the "sunet-cdn-agent-"
prefix when discarding old configs, so for now the initial config named
"boot" which is created at startup will never be discarded.