Add WIP l4lb node management
* Fetch l4lb config from the api/v1/l4lb-node-configs endpoint.
* Manage IP address setup in the l4lb namespace for the allocated
  service IP addresses. This is done by filling in
  "netns-sunet-cdn-agent.json" and calling "systemctl restart
  sunet-l4lb-namespace.service". For this to work you need to have the
  machine prepared with the "cdn::l4lb" class from
  https://platform.sunet.se/sunet-cdn/cdn-ops
* Manage ipvsadm rules for sending traffic from the l4lb machine to
  backend cache nodes. This is currently a WIP and right now
  expected modifications are printed rather than executed.

The ipvsadm management became unwieldy enough to be broken out into a
separate package. It is unfortunate that ipvsadm --restore is not
natively taking care of updating the current loaded rules to what
exactly is in the restore file created with ipvsadm --save (it just
executes the lines one by one, and does not cleanup any loaded rules
that have been removed from the file).