Support cleanup of orgs and services on cache node
If a service is removed we stop related containers, remove systemd unit
file config and clean up related directory structures on disk. If the
last service for an org is removed this results in cleaning up the org
directory as well.

Add some minimal helper functions for building paths to the different
directories that are used both for creation and deletion to more easily
make sure both operations look at the same paths.