Start using relative VCL filename
Since we removed the default "+unsafe_path" from "vcc_feature" the
dynamic VCL reloads done via varnishadm broke the following error:
```
VCL filename '/service-versions/active/varnish/default.vcl' is unsafe.
```

So make this work by instead configuring the vcl_path and referring to
the config file without slashes. This also included the change of
renaming default.vcl to sunet-cdn.vcl to make it more clear that it is
not the /etc/varnish/default.vcl present in the container that is being
used.

Start marking vcl_path readonly as well since there should be no
circumstances this needs to be modified at runtime either.