React to configuration changes for active versions
Before this we would only notify containers to reload config if the
active symlink was modified. Because we might modify the templates used
for generating configurations this can affect an active version even if
the symlink does not change.

Now we also react if the active config is modified in-place.